feat: add Specialty model and alter Doctor model
This commit adds a new model called Specialty and makes alterations to the Doctor model. The Specialty model includes fields for spec_id and spec_name, while the Doctor model has been updated to include a foreign key relationship with the Specialty model. This change allows doctors to be associated with specific specialties.

The commit also includes a migration file for these changes.

Refactor user forms and views

This commit refactors the user forms and views. It improves the code structure and organization, making it more maintainable and easier to understand.

Renamed email.py to utils.pt

This commit renames the file email.py to utils.pt. This change was made to improve consistency and clarity in the codebase.

fix: forignkey models import

This commit fixes an issue with the import of foreign key models. It ensures that the correct models are imported and used in the code.

fix: signup and login problems

This commit fixes problems related to the signup and login functionality. It addresses issues that were causing errors and improves the overall user experience.

feat: add doctor form

This commit adds a new form for doctors. The form allows users to input information about doctors, such as their name, specialty, and availability. This feature enhances the functionality of the application and improves the user experience.

fix: redirect in create_wallet view

This commit fixes a redirect issue in the create_wallet view. It ensures that users are redirected to the correct page after creating a wallet.